fix can not delete if not select data

+++ lib/app/modules/station/manageStation/views/manage_station_view.dart
@@ -233,41 +233,43 @@ class ManageStationDetail extends StatelessWidget {
                         IconButton(
                           icon: const Icon(Icons.delete_sharp),
                           onPressed: () async {
-                            final result = await Get.dialog(
-                              AlertDialog(
-                                content: const CustomText(
-                                  text: "ยืนยันการลบข้อมูล ?",
-                                  scale: 1.2,
-                                ),
-                                actions: [
-                                  TextButton(
-                                    child: const CustomText(
-                                      text: "ยืนยัน",
-                                      scale: 1.2,
-                                      color: Colors.red,
-                                    ),
-                                    onPressed: () => Get.back(result: 'Y'),
+                            if (controller.selectedId > 0) {
+                              final result = await Get.dialog(
+                                AlertDialog(
+                                  content: const CustomText(
+                                    text: "ยืนยันการลบข้อมูล ?",
+                                    scale: 1.2,
                                   ),
-                                  TextButton(
-                                    child: const CustomText(
-                                      text: "ปิด",
-                                      scale: 1.2,
-                                      color: Colors.green,
+                                  actions: [
+                                    TextButton(
+                                      child: const CustomText(
+                                        text: "ยืนยัน",
+                                        scale: 1.2,
+                                        color: Colors.red,
+                                      ),
+                                      onPressed: () => Get.back(result: 'Y'),
                                     ),
-                                    onPressed: () => Get.back(result: "N"),
-                                  ),
-                                ],
-                              ),
-                            );
-                            if (result == 'Y') {
-                              Get.dialog(
-                                const Center(
-                                  child: CircularProgressIndicator(),
+                                    TextButton(
+                                      child: const CustomText(
+                                        text: "ปิด",
+                                        scale: 1.2,
+                                        color: Colors.green,
+                                      ),
+                                      onPressed: () => Get.back(result: "N"),
+                                    ),
+                                  ],
                                 ),
-                                barrierDismissible: false,
                               );
-                              await controller.delete();
-                              Get.back();
+                              if (result == 'Y') {
+                                Get.dialog(
+                                  const Center(
+                                    child: CircularProgressIndicator(),
+                                  ),
+                                  barrierDismissible: false,
+                                );
+                                await controller.delete();
+                                Get.back();
+                              }
                             }
                           },
                         ),
